api_legal/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-cors-headers
django-filter
